fix mismatched ints

+++ internal/lib/dump_waf_darwin.go
@@ -86,7 +86,7 @@ func DumpEmbeddedWAF() (_ string, closer func() error, err error) {
 			return "", nil, fmt.Errorf("error rlocking file %s: %w", path, err)
 		}
 
-		if fileinfo, err := fp.Stat(); err != nil && fileinfo.Size() < len(libddwaf) {
+		if fileinfo, err := fp.Stat(); err != nil && fileinfo.Size() < int64(len(libddwaf)) {
 			// The file is not written yet, so we need to unlock it and try again.
 			if err = unlock(fp); err != nil {
 				return "", nil, fmt.Errorf("error unlocking file %s: %w", path, err)
